refactor B)
